backend/requirements.txt
# Django
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Authentication & Security
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0
django-otp>=1.5,<2.0
qrcode>=7.4,<8.0

# Permissions
django-guardian>=2.4,<3.0

# Database
psycopg[binary]>=3.1,<4.0

# Search
meilisearch>=0.31,<1.0

# Async tasks
celery>=5.4,<6.0
django-celery-beat>=2.6,<3.0
redis>=5.0,<6.0

# File handling
openpyxl>=3.1,<4.0
lxml>=5.0,<6.0

# Production server
gunicorn>=22.0,<23.0

# Utilities
python-decouple>=3.8,<4.0
django-extensions>=3.2,<4.0
Pillow>=10.0,<11.0
